feat(gormx): Simplify gormx and make helpers private
